Fix a bug introduced by SocketIO not having consistent ids
During the time of the initial release of Peer Calls, the server and
client sockets used to have different namespaces - one side had `/#`
prepended to the name, whereas the other did not, so I had to check for
this in the code.

This was fixed since the release of [SocketIO v1.5.0][1], thus breaking the
compatibility with PeerCalls. Any new `npm install` would break this
because of the way it works - it tries to install the latest available
"compatible" release.

As of this commit, the SocketIO version is locked to v1.6.0.

[1]: https://github.com/socketio/socket.io/releases/tag/1.5.0