Fix too many videos back-and-forth
This took a long time to figure out. The bottom line is that
AddTransceiverFromTrack with Direction sendonly should be used instead
of AddTrack, and AddTransceiverFromKind with Direction recvonly when
expecting to receive a track.

Otherwise, the peer connection might go back and forth and trying to add
a new track...