Use a channel for signalling peer connection close
There's a bug that needs fixing. Currently, if a user hangs up, stays on
the same page and tries to reconnect immmediatelly, the server
connection might not have timed out yet. Since the user will still have
the same clientID, a part of the cleanup will happen on a new peer
connection instead of the old one.

Also, there's probably a race condition with signaller being set to nil