Refactor channels to never panic on error
Previous implementation assumed that starting writing to an unbuffered
channel would not panic as long as a check was made before writing to
channel. However, an writing to an unbuffered channel will wait until
the value is read on the other side. If such a channel was closed before
the other side read (before the value was sent), it writing could panic.
This commit adds a second "close" channel which is only used for
signalling the closing.

The channel write methods will now acquire a RLock, check if a channel
was closed. If the channel was closed, it will use a nil channel which
blocks indefinitely and try to read from closeChannel. For example:

type Item struct {
  channel      chan string
  closeChannel chan struct{}
  closed       bool
  mu           sync.RWMutex
}

func (s *Item) Close() {
  s.once.Do(func() {
    close(s.closeChannel)

    s.mu.Lock()
    defer s.mu.Unlock()

    close(s.channel)
    s.closed = true
  })
}

func (s *Item) Write(payload string) {
  s.mu.RLock()

  go func() {
    defer s.mu.RUnlock()

    ch := s.channel
    if s.closed {
      ch = nil
    }

    select {
    case ch <- payload:
      // successfully sent
    case <-s.closeChannel)
    }
  }()
}

Also refactor ws.Client. Client.Write() fn should be used instead of
WriteChannel. The Client.Subscribe() function has been simplified, and
Client.Err() method added to check for errors.

Also add test for establishing a peer connection when using P2S (SFU).