Add received track immediately when not muted
When a track is received with muted=false, add it immediately. This
fixes an issue in Safari where the video wouldn't play when a new track
was added from the peer. Safari doesn't seem to automatically trigger
the onunmute event so tracks received with muted=false have to be added
as soon as they are received.