Prevent RedisAdapter.Close() deadlock in tests
The Close() fn acquires the clientsMu lock and it eventually calls
stop() function which tries cancels the subscription context. However,
if the subscription was handling a message around that time, it could
result in a deadlock because it eventually tries to acquire the
clientsMu lock.

The fix moves the clientsMu.Lock() call later in close, after the
subscription is already closed.