Add Transport abstraction
This should help with communication across multiple Peer Calls nodes.

The PeerTransport is currently the only implementation, and it is used
to add/remove tracks and send raw rtp/rtcp packets to a peer connection.

The next one will be ServerTransport which should allow the servers to
exchange these packets without the overhead of a webrtc peer connection
handshaking since the servers will have IPs and ports exposed.