Add ControlTransport
The idea is that SCTP StreamID #0 will be used as a control transport
for the whole association.

This brings several benefits:

1. No more need for a separate channel per room for metadata transport
2. TODO: Add ability for negotiation stream IDs

The ControlTransport provides an adapter for ServerMetadataTransport
which will forward track related events per room to a specific
transport.

Something to think about: not sure if we still need to add room related
metadata in TracksManager to track events since this will also be
handled by the adapter.