Update factory but realize this is a failed attempt
There are too many problems with the design where we _first_ initiate a
sctp connection, and _then_ have a stringmux that splits the data by
rooms. This would mean we'd need to manage three separate stringmuxes
for each part of the server transport (data, metadata and media), and
that's just too much because we wouldn't be able to create a transport
once a first packet is received.

Actually, scratch that. This might just work - we can still call
stringmux.GetConn() for the other two as soon as the first conn is
received.