Add better logging to metadata_transport.go
Status: udptransport2 tests are failing (flaky). The reason is the fact
that the new init event sometimes triggers a new connection even after
the first one has been closed.

See TODO.md for more ideas on how to fix this. Also see the state
diagram in transport_state.go. If factory is redesigned to use a control
SCTP stream channel for initiating transports, it might be able to make
use of transportStateTracker for this, and only initiate a transfer once
both local and remote rooms are ready for sending and receiving (e.g.
both have peers in rooms).