docker: Use version as external arg
The problem was that `git describe` would always result with a `-dirty`
suffix because we wouldn't include `Dockerfile` and `build/.gitkeep` in
the docker build context.

We don't really want to include `Dockerfile` because any change to it
would invalidate the build cache, and we don't want that.

So instead of using `git describe` within the docker build context, we
pass the VERSION using `--build-arg` parameter.