Fix device dropdown on some Android devices
Some android devices do not support multiple calls to `getUserMedia`
with different `deviceId`s set, possibly because of the limitations in
the USB bus size when large sensors are used.

We can't call `stop` because that would invalidate the whole WebRTC
track; Want to be able to call `replaceTrack` on the peer connection to
quickly change the camera without renegotation, so as a workaround we'll
create a dummy black canvas video track and replace the video track.
When we replace this track, the existing real video track will be
stopped and we can safely call `getUserMedia` again.

Tested and verified on latest versions of Chrome and Firefox on Samsung
Galaxy A52s.