Implement a better video grid (#251)
The new video grid supports fixed aspect ratios and will not resize the
video like it was done previously. The new layout is enabled
only when there are more than three participants. Users have a choice of
picking the three modes in the settings: AUTO (the new default), LEGACY
(the old default) and ASPECT (the new mode).

We can now also have a menu on the top videos toolbar that flips over
the edge of the parent container with defined height. On the other side,
this position is only defined once so if a window is resized, the
fixed-positioned dropdown will stay in place. We have a dropdown
backdrop so it's not the end of the world, but it's a compromise.

Fix an issue with toolbar on Firefox for Android due to tooltips being
too wide, for some reason it added extra margins on the bottom/right.