TMP: handle video dimensions, next step: use it to calculate the Frame
Since videos may have different dimensions and orientation (e.g.
portrait/landscape), we can and calculate the Frame sizes dynamically.

We just need to count the amount of unique aspect ratios we have and
then we extend the existing formula from:

    N * x * y = <= screenW * screenH

    A * x1 * y + B * x2 * y + C * x3 * y + ... <= screenW * screenH

Since we know the aspect ratios, or the link between each x/y, we can
reduce the above formula to some just y and calculate the y; after
that it's easy to calculate the x1, x2 and x3...