Fix: deadlock in server.Signaller.Close
There was a deadlock in `Signaller` if `Signaller.Close` if an ICE
candidate (signal) was being handled at the same time, since both
`onSignal` and `Close` would try to acquire the same `signalMu` lock:

    Signaller.Close() called
         Signaller.signalMu.lock() called
              Signaller.peerConnection.Close() called
    |----|----|------------------------------->
    t0   t1   t2                              t

If `webrtc.PeerConnection` generates an ICE candidate just after the
`signalMu.lock()` is acquired, but before
`Signaller.peerConnection.Close()` had been called (after t1, but before
t2), the peer connection would call the registered callback, `onSignal`,
which would try to acquire the same lock. Because the
`webrtc.PeerConnection.Close` waits until `ICETransport` shuts down, it
would deadlock indefinitely.

The only reason there's a lock in the first place is so we can safely
close the `signalChannel` after the peer connection is closed. In theory
we should be able to just close the `signalChannel` after the
`webrtc.PeerConnection` is closed, but I remember dealing with some edge
cases a couple of years ago and it didn't quite work. The issue is that
some handlers are called in a new goroutine and there can still be a
stale goroutine _after_ the peer connection is closed.

I used this test to assert this fix actually works (placed in
`server/sfu_test.go`), but it would occasionally leave some goroutines
running because they didn't have the chance to proprely tear down so I
didn't leave the test:

    func TestSFU_PeerConnection_CloseImmediately(t *testing.T) {
      log := test.NewLogger()

      defer goleak.VerifyNone(t)

      newAdapter := server.NewAdapterFactory(log, server.StoreConfig{})
      defer newAdapter.Close()

      rooms := server.NewAdapterRoomManager(newAdapter.NewAdapter)
      srv, wsBaseURL := setupSFUServer(rooms, false)
      defer srv.Close()

      ctx, cancel := context.WithTimeout(context.Background(), timeout)
      defer cancel()

      peerCtx := createPeerConnection(t, ctx, wsBaseURL+roomName.String()+"/"+clientID.String(), clientID)
      peerCtx.close()

      // Allow the remaining goroutines to terminate.
      runtime.Gosched()
    }